feat: [#55] new option --dvc-remote for workflow command
That option allows you to overwrite the default DVC remote used in the
Gold Image Procesing command.
