+++ src/nautilus_librarian/mods/dvc/domain/utils.py
@@ -99,6 +99,14 @@ def extract_added_files_from_dvc_diff(dvc_diff):
     return [(path_object["path"]) for path_object in data["added"]]
 
 
+def dvc_default_remote(git_repo_dir):
+    """
+    It returns the default remote for the dvc repo.
+    """
+    output = execute_console_command("dvc remote default --project", cwd=git_repo_dir)
+    return output.strip()
+
+
 @deprecated(reason="use DvcApiWrapper class")
 def dvc_add(filepath, git_repo_dir):
     """
